requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-core>=0.1.0
langgraph>=0.0.20
langgraph-cli[inmem]>=0.2.6
fastapi>=0.115.0
openai>=1.0.0
python-dotenv>=1.0.0
langsmith>=0.1.0
reportlab>=4.2.0
